Bugfix #7, incorrect top n% on fit_tests().
It's a minor bug. When defining the top n% of the samples to fit the
exponweib.pdf(), it was considering the total N that included non valid
data.

A sample can be valid, but not possible to evaluate in a specific test,
like a place with a climatology build with only one historical sample.
In that case, I do not consider the climatology result, so there is a
sample that could be flagged valid by other tests, but do not have a
climatology test result.