Tuple/list safe (#52)
Changes in the last years restricted some procedures from running with tuples. Guarantee that it can be used again with bare tuples/lists.

* refactor: Avoid bare except

* test: Testing rate of change

* feat: rate of change with tuple

rate_of_change and RateOfChange can process a tuple of data.

* feat: global range with tuple

* feat: spike with tuple

* pref: Don't need to consider masked array anymore

* feat: gradient with tuple

* feat: gradient|spike depth conditional with tuple

* feat: profile envelop with tuple

* feat: WOA comparison test with tuple

* feat: Digit roll over test with tuple

* flag as numpy.array

* Estimate the normalized tukey53H only when l is given

* feat: Verify type and dtype of flags

Procedures on the family QCCheck should return a flag component that
must be a numpy.array dtype int8. Confirm that when testing.

* style: Cleaning, mostly removing useless code

* Missing __init__ in tests sub-modules

* Fixing path to load submodule data for testing

* Improving comparison in tests for multiple input arguments

The feature density_step is one example of a function that requires more
than one input argument. The compare procedure was expanded to allow
such a scenario with multiple inputs.

* feat: Density inversion test with tuple

* feat: Location at sea test with tuple

* feat: Valid Geolocation test with tuple

* refactor: Moved extract_coordinates to utils

The function extract_coordinates() might be used in more places.

* fix: DensityInversion create feature conditional to gsw

The densitystep() fails if missing gsw,  and DensityInversion returns
an empty feature if missing gsw.

* density inversion test conditional to gsw availability

* docs: Adding some notes.

* docs: Fixing wrong documentation on Morello's fuzzy logic

* feat: Adding a LocationAtSea test inland