Flexible WOA comparison (#57)
* WIP: different date sources

* feat: WOA comparison features are regular arrays

Conform with new standard for CoTeDe of all features as regular arrays.

Note, for number of samples, which is an integer type, I'm using -1 to
keep as an integer. I might regret on this decision.

* refactor: woa_normbias() in charge of WOA comparison

woa_normbias() was several years old, and originally would do the whole
WOA QC comparison. The standard QC classes were developed and adopted
later. This commit was a step towards bringing the WOA comparison to the
current standard, where woa_normbias() takes care of defining the
features, i.e. obtaining the WOA refernce values as well as the bias and
normbias.

* feat: extract_time(), search for time information

Different datasets with different structure might also follow different
vocabularies, thus extract_time search on different places for the time
to associate with the measurements.

* fix: typo

* style: PEP8

* fix: moving test_flex_time into tests

* feat: Extract DOY from several possible inputs

utils.day_of_year() is a WIP but it is fine for now. It can obtain the
DOY as integer from several possibilities of inputs. Once I normalize
other parts of the code I will come back here to finish and simplify
this function.

* feat: Extract depth from several possible standards

* refactor: WOA_NormBias uses woa_normbias to create features

* feat: More validation tests for woa_normbias

* fix: Updating required version of OceansDB

There was a bug on OceansDB for invalid locations such as on land.

* refactor: Fails if can't obtain geolocation

Before the approach was to create an empty features dictionary, but now
that I'm using woa_normbias it breaks if can't obtain lat lon.

* feat: More logging information

* refactor: minimum samples for WOA comparison

Repositioning minimum samples control in WOA_NormBias class

* refactor: flag 0 if woa_normbias fails

* docs: typo

* docs: Comment on explicit target variables

* fix: Wrong path to support module - data

* fix: Require generic Anom. Detection to skip test

This new approach on WOA comparison is incompatible with old Anomaly
Detection approach. The new structure, in a parallel branch will be
merged soon.

* Bump version: 0.23.2 → 0.23.3

* doc: Clarifying warning message