Generic procedure (#56)
Updating public CoTeDe to take advantage of the standard QC classes in a generalized procedure.

* Catalog of procedures

* cfg defining explicit procedures

* refactor: Moving list of procedures away from qc.py

ProfileQC should be as generic as possible. With a cfg, qctests should
return the list of procedures to be applied.

* fix: pressure_increasing using MonotonicZ

MonotonicZ has been available in a while, and the increasing pressure
test should use that, with explicity definition to pressure. Here I
removed an outdated warning.

* refactor: Placeholder for varname at QCCheck

This allows to generalize the use of QCCheck, since it ignores a varname
if that is given.

* feat: QCCheck.test now raises NotImplementedError

QCCheck is the fundamental QC object and it is expected that the child
objects defines its own test procedure.

* refactor: Moving QCCheck & QCChecVar around