Fuzzylogic (#60)
Refactoring the Fuzzy Logic procedures. The code was simplified while generalized. There is a new option to allow an incomplete record of features, instead of the default strict requirement on all features. An important fix that allows returning flag zero is lacking sufficient information to estimate the uncertainty. Significant improvements on the validation tests.

* refactor: Conforming fuzzy logic with the standard procedure

Fuzzy logic was implemented several years ago, and the way CoTeDe QC
evolved. The procedure for fuzzy logic is now closer to the standard
approach.

* style: PEP8

* feat: Improving flexibility of fuzzy memberships

Improving Fuzzy membership functions to handle better handle more types
and shapes of inputs, as well as special types such as NaN & Inf.

* style: PEP8

* doc: Improving documentation on fuzzy membership

* feat: fuzzify's optional argument: require

Refactor fuzzify adding the option require to control the requirement on
all features must be valid or at least one valid feature.

* refactor: Expanding fuzzyfy arguments

* feat: QCCheckVar accepts explicit attrs and pass it to QCCheck

* feat: Extending tested values for memberships

* feat: Expanding fuzzy_uncertainty arguments

* refactor: cleanning fuzzyfy test

* fix: Return arrays of NaN when fuzzify with all NaN features

* feat: Testing fuzzy_uncertainty with several input types

* fix: fuzzy_uncertainty response to NaN inputs

Note that it is possible to use the argument require to alternate the
response between requiring all or any valid feature.

* reafctor: Updating fuzzylogic(), the feature estimator

* fix: Missing packages on CI

* fix: Missing __init__.py

* feat: Validate compound features types

* fix: Missing FuzzyLogic procedure object

* fix: Updating Morello2014 execpted flags

The fuzzy unceartainty was updated to require (default choice) all
features, thus spike & gradient limits the unceartainty estimate in the
bounds.